Bugfix map file sources
