+++ src/Kart/LeafletTangram/scene/visualisering/gradient.js
@@ -1,3 +1,5 @@
+import sysconfig from '../../../../config'
+
 function drawAll(drawArgs) {
   const layer = {
     data: { source: drawArgs.kode },
@@ -34,23 +36,13 @@ function lagStyle({ filterMin, filterMax }) {
 }
 
 function lagSource(kode, bbox, zoom) {
-  if (!bbox || !zoom) {
-    console.warn(`No map data for ${kode}`)
-  }
-  const source = {
-    type: 'Raster',
-    url: `https://{s}.artsdatabanken.no/gradient/${kode}/{z}/{x}/{y}`,
-    url_subdomains: ['nintest', 'rover'],
-  }
+  const source = sysconfig.createTileSource(
+    `gradient/${kode}`,
+    'Raster',
+    zoom,
+    bbox
+  )
 
-  if (bbox) {
-    const [ll, ur] = bbox
-    source.bounds = [ll[1], ll[0], ur[1], ur[0]]
-  }
-  if (zoom) {
-    source.min_zoom = zoom[0]
-    source.max_zoom = zoom[1]
-  }
   return source
 }
