+++ src/Kart/LeafletTangram/scene/visualisering/polygon.js
@@ -1,4 +1,5 @@
 import tinycolor from 'tinycolor2'
+import sysconfig from '../../../../config'
 
 function drawAll({
   kode,
@@ -75,23 +76,7 @@ function draw(args) {
 }
 
 function lagSource(kode, bbox, zoom) {
-  if (!bbox || !zoom) {
-    console.warn(`No map data for ${kode}`)
-  }
-  const source = {
-    type: 'MVT',
-    url: `https://{s}.artsdatabanken.no/polygon/${kode}/{z}/{x}/{y}`,
-    url_subdomains: ['nintest', 'rover'],
-  }
-  if (bbox) {
-    const [ll, ur] = bbox
-    source.bounds = [ll[1], ll[0], ur[1], ur[0]]
-  }
-  if (zoom) {
-    source.min_zoom = zoom[0]
-    source.max_zoom = zoom[1]
-  }
-  return source
+  return sysconfig.createTileSource(`polygon/${kode}`, 'MVT', zoom, bbox)
 }
 
 export default { drawAll, lagSource }
