+++ src/config.js
@@ -6,6 +6,27 @@ class config {
     comboSøk: false,
   }
 
+  static createTileSource(relativePath, type, zoom, bbox) {
+    const source = {
+      type: type,
+      url: `https://{s}.artsdatabanken.no/${relativePath}/{z}/{x}/{y}`,
+      url_subdomains: ['nintest'],
+    }
+    if (!bbox || !zoom) {
+      console.warn(`No map extents for ${relativePath}`)
+    }
+
+    if (bbox) {
+      const [ll, ur] = bbox
+      source.bounds = [ll[1], ll[0], ur[1], ur[0]]
+    }
+    if (zoom) {
+      source.min_zoom = zoom[0]
+      source.max_zoom = zoom[1]
+    }
+    return source
+  }
+
   static getFotoOmslag(
     kode: string,
     width: number = 408,
